MPEG4 Enc flush port workaround
This is a temporary workaround that allows the MPEG4 720p test case to be unblocked. OMX should return a signal for the port flushing event. The Aricent implementation does not do that so even though the resources are released, the commmand prompt is not returned. The functionality of the component is not broken due to this modification.